+++ java/openmldb-jdbc/pom.xml
@@ -61,6 +61,11 @@
             <artifactId>snappy-java</artifactId>
             <version>1.1.7.2</version>
         </dependency>
+        <dependency>
+            <groupId>com.github.ben-manes.caffeine</groupId>
+            <artifactId>caffeine</artifactId>
+            <version>2.9.3</version>
+        </dependency>
 
         <!-- https://mvnrepository.com/artifact/org.testng/testng -->
         <dependency>
